+++ lib/modules/app_module.dart
@@ -1,13 +1,13 @@
-import 'package:app_test/modules/login/external/datasources/login_datasource_impl.dart';
-import 'package:app_test/modules/login/presenter/pages/login_page.dart';
+import 'package:app_test/modules/authentication/external/datasources/login_datasource_impl.dart';
+import 'package:app_test/modules/authentication/presenter/pages/login_page.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter/material.dart';
 
 import 'app_widget.dart';
-import 'login/data/repositories/login_repository_impl.dart';
-import 'login/domain/repositories/login_repository.dart';
-import 'login/domain/usecases/result_login_usecase.dart';
+import 'authentication/data/repositories/login_repository_impl.dart';
+import 'authentication/domain/repositories/login_repository.dart';
+import 'authentication/domain/usecases/result_login_usecase.dart';
 
 class AppModule extends Module {
   @override
@@ -20,6 +20,6 @@ class AppModule extends Module {
 
   @override
   final List<ModularRoute> routes = [
-    ChildRoute(Modular.initialRoute, child: (_, __) => const LoginPage()),
+    ChildRoute(Modular.initialRoute, child: (_, __) => LoginPage()),
   ];
 }
