+++ lib/modules/authentication/domain/entities/result_login_entity.dart
