+++ lib/modules/authentication/domain/errors/errors.dart
