+++ lib/modules/authentication/presenter/pages/login_page.dart
@@ -1,0 +1,158 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:validators/validators.dart';
+import '../../../../core/constants/app_constants.dart';
+import '../cubit/auth_cubit.dart';
+import '../cubit/auth_state.dart';
+import '../widgets/bottom.dart';
+import '../widgets/custom_check.dart';
+import '../widgets/custom_textfield.dart';
+import '../widgets/default_btn.dart';
+import '../widgets/header.dart';
+import '../widgets/link.dart';
+import '../widgets/logo.dart';
+
+class LoginPage extends StatefulWidget {
+  LoginPage({Key? key}) : super(key: key);
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  final TextEditingController _emailController = TextEditingController();
+
+  final TextEditingController _passwordController = TextEditingController();
+
+  final _formKey = GlobalKey<FormState>();
+
+  String? validatePassword(String? value) {
+    if (value == null || value.isEmpty) {
+      return AppLocalizations.of(context)!.pleaseInsertYourPassword;
+    }
+    return null;
+  }
+
+  String? validateEmail(String? value) {
+    if (value == null || value.isEmpty) {
+      return AppLocalizations.of(context)!.pleaseInsertYourEmail;
+    } else if (!isEmail(value)) {
+      return AppLocalizations.of(context)!.pleaseInsertValidEmail;
+    }
+    return null;
+  }
+
+  void validate(BuildContext context) {
+    if (_formKey.currentState?.validate() ?? false) {
+      final email = _emailController.text;
+      final password = _passwordController.text;
+      context.read<AuthCubit>().login(email, password);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    ScreenUtil.init(
+      context,
+      designSize: const Size(375, 812),
+      minTextAdapt: true,
+      splitScreenMode: true,
+    );
+
+    return Scaffold(
+      body: BlocProvider(
+        create: (context) => AuthCubit(),
+        child: BlocListener<AuthCubit, AuthState>(
+          listener: (context, state) {
+            if (state is AuthAuthenticated) {
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(content: Text('Login bem-sucedido!')),
+              );
+            } else if (state is AuthError) {
+              String message = '';
+              if (state.message == 'invalidCredentials') {
+                message = AppLocalizations.of(context)!.invalidCredentials;
+              } else {
+                message = AppLocalizations.of(context)!.loginError;
+              }
+
+              ScaffoldMessenger.of(context).showSnackBar(
+                SnackBar(content: Text(message)),
+              );
+            }
+          },
+          child: Padding(
+            padding: EdgeInsets.symmetric(horizontal: 20.w),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                SizedBox(height: 100.h),
+                const Logo(path: AppConstants.logo_path),
+                SizedBox(height: 30.h),
+                Header(
+                  title: AppLocalizations.of(context)!.welcomeBack,
+                  subtitle:
+                      AppLocalizations.of(context)!.signInToAccessYourAccount,
+                ),
+                SizedBox(height: 40.h),
+                Form(
+                  key: _formKey,
+                  child: Column(
+                    children: [
+                      CustomTextField(
+                        controller: _emailController,
+                        validator: validateEmail,
+                        label: AppLocalizations.of(context)!.enterYourEmail,
+                        icon: const Icon(Icons.email),
+                      ),
+                      SizedBox(height: 20.h),
+                      CustomTextField(
+                        controller: _passwordController,
+                        validator: validatePassword,
+                        label: AppLocalizations.of(context)!.password,
+                        obscureText: true,
+                        icon: const Icon(Icons.lock),
+                      ),
+                    ],
+                  ),
+                ),
+                SizedBox(height: 10.h),
+                Row(
+                  children: [
+                    CustomCheck(
+                        title: AppLocalizations.of(context)!.rememberMe),
+                    const Spacer(),
+                    Link(
+                      title: AppLocalizations.of(context)!.forgetPassword,
+                      url: '',
+                    ),
+                  ],
+                ),
+                const Spacer(),
+                BlocBuilder<AuthCubit, AuthState>(
+                  builder: (context, state) {
+                    return DefaultBtn(
+                      title: AppLocalizations.of(context)!.next,
+                      icon: const Icon(Icons.arrow_forward_ios,
+                          color: Colors.white, size: 18),
+                      onPressed: () => validate(context),
+                      isLoading: state is AuthLoading,
+                    );
+                  },
+                ),
+                SizedBox(height: 20.h),
+                Bottom(
+                  title: AppLocalizations.of(context)!.newMember,
+                  textLink: AppLocalizations.of(context)!.registerNow,
+                ),
+                SizedBox(height: 30.h),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
