+++ lib/modules/authentication/presenter/widgets/custom_textfield.dart
@@ -1,0 +1,35 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class CustomTextField extends StatelessWidget {
+  final String label;
+  final Icon icon;
+  final bool? obscureText;
+  final TextEditingController controller;
+  final String? Function(String?)? validator;
+  const CustomTextField({
+    Key? key,
+    required this.label,
+    required this.icon,
+    this.obscureText,
+    required this.controller,
+    this.validator,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      obscureText: obscureText ?? false,
+      controller: controller,
+      validator: validator,
+      decoration: InputDecoration(
+        labelText: label,
+        prefixIcon: icon,
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10.r),
+        ),
+      ),
+    );
+  }
+}
