+++ lib/modules/authentication/presenter/widgets/default_btn.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class DefaultBtn extends StatelessWidget {
+  final String title;
+  final Icon icon;
+  final VoidCallback onPressed;
+  final bool isLoading;
+
+  const DefaultBtn({
+    Key? key,
+    required this.title,
+    required this.icon,
+    required this.onPressed,
+    this.isLoading = false,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      style: ElevatedButton.styleFrom(
+        minimumSize: Size(double.infinity, 50.h),
+        backgroundColor: const Color(0xFFFF3951),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(10.r),
+        ),
+      ),
+      onPressed: isLoading ? null : onPressed,
+      child: isLoading
+          ? const CircularProgressIndicator(
+              valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+            )
+          : Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text(
+                  title,
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 16.sp,
+                    fontWeight: FontWeight.w600,
+                  ),
+                ),
+                SizedBox(width: 10.w),
+                icon,
+              ],
+            ),
+    );
+  }
+}
