+++ lib/modules/authentication/presenter/widgets/logo.dart
