--- lib/modules/login/domain/repositories/login_repository.dart
@@ -1,8 +1,0 @@
-import 'package:app_test/modules/login/data/models/login_model.dart';
-import 'package:app_test/modules/login/domain/entities/result_login_entity.dart';
-import 'package:app_test/modules/login/domain/errors/errors.dart';
-import 'package:dartz/dartz.dart';
-
-abstract class LoginRepository {
-  Future<Either<LoginFailure, ResultLoginEntity>> login(LoginModel loginModel);
-}
