--- lib/modules/login/presenter/widgets/custom_textfield.dart
@@ -1,28 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-class CustomTextField extends StatelessWidget {
-  final String label;
-  final Icon icon;
-  final bool? obscureText;
-  const CustomTextField({
-    Key? key,
-    required this.label,
-    required this.icon,
-    this.obscureText,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      obscureText: obscureText ?? false,
-      decoration: InputDecoration(
-        labelText: label,
-        prefixIcon: icon,
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10.r),
-        ),
-      ),
-    );
-  }
-}
